unify into outputs directory
